refactor scattergl and splom - rename style > edit_style
