+++ src/traces/scattergl/edit_style.js
